isolate opennest requirements
